Revert "Merge pull request #102217 from tkashem/client-go-refactor-retry"
This reverts commit a8985871f9d8e069efdc6b82f41dbed8a39389ee, reversing
changes made to 77937b1e8eb10134bb989680520d901f1b4ac076.
